till logged out

+++ src/utils/cloudinary.js
@@ -1,24 +1,29 @@
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
+
+
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
   api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
-const uploadOnCloudinary= async ( localFilePath) =>{
+console.log("ENV CHECK:", process.env.CLOUDINARY_CLOUD_NAME, process.env.CLOUDINARY_API_KEY);
+const uploadOnCloudinary = async (localFilePath) => {
+    if (!localFilePath) return null;
+    const normalizedPath = localFilePath.replace(/\\/g, "/");
+
     try {
-        if(!localFilePath) return null
-       const response =await cloudinary.uploader.upload(localFilePath,{
-            resource_type:"auto"
-        })
-        console.log("file is uploaded on cloudinary",response.url);
+        const response = await cloudinary.uploader.upload(normalizedPath, {
+            resource_type: "auto"
+        });
+       fs.unlinkSync(localFilePath)
         return response;
-        
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the  upload operation got failed
+        console.dir(error, { depth: null });
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
         return null;
     }
 }
 
-export {uploadOnCloudinary}
+export { uploadOnCloudinary }
